+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 - Added Import and Export of bookmarks as described in Issue #36
 - Added Error Reports as described in Issue #37
+- Added Pinyin Converter as described in Issue #23
+- Added Character Converter as described in Issue #22
 
 ## [0.1.2] - 2016-04-18
 ### Fixed